// src/untils/enums.ts
export enum UserRole {
  ADMIN = 'ADMIN',
  EMPLOYEE = 'EMPLOYEE',
}

export enum GenderType {
  MALE = 'Male',
  FEMALE = 'Female',
}

export enum StatusType {
  TODO = 'ToDo',
  DOING = 'Doing',
  DONE = 'Done',
}

export enum CouponStatus {
  ACTIVE = 'active',
  INACTIVE = 'inactive',
}

export enum PaymentMethod {
  CARD = 'card',
  WALLET = 'wallet',
  CASH = 'cash',
}

export enum PaymentStatus {
  PENDING = 'PENDING',
  PAID = 'PAID',
  FAILED = 'FAILED',
  CANCELLED = 'CANCELLED',
  REFUNDED = 'REFUNDED',
}


export enum QuestionType {
  MCQ = 'mcq',       
  TRUE_FALSE = 'true_false', 
}

export enum EnrollmentType {
  INSTRUCTOR = 'INSTRUCTOR',
  STUDENT = 'STUDENT',
}

export enum ConversationStatus {
  OPEN = 'OPEN',
  HUMAN = 'HUMAN',
  CLOSED = 'CLOSED',
}

export enum MessageRole {
  USER = 'USER',
  ASSISTANT = 'ASSISTANT',
  HUMAN = 'HUMAN',
}